Removes unnecessary code in tests and case for general and custom error handling
