+++ README.md
@@ -1,10 +1,8 @@
-![image](https://img.shields.io/badge/JavaScript-323330?style=for-the-badge&logo=javascript&logoColor=F7DF1E)
-![image](https://img.shields.io/badge/npm-CB3837?style=for-the-badge&logo=npm&logoColor=white)
-
+![test workflow](https://github.com/github/docs/actions/workflows/test.yml/badge.svg)
 
 # Appyx-Comm
 
-![test workflow](https://github.com/github/docs/actions/workflows/test.yml/badge.svg)
+![image](https://img.shields.io/badge/JavaScript-323330?style=for-the-badge&logo=javascript&logoColor=F7DF1E) ![image](https://img.shields.io/badge/npm-CB3837?style=for-the-badge&logo=npm&logoColor=white)
 
 Communication package for javascript.
 
@@ -16,12 +14,12 @@ $ npm i @eryxcoop/appyx-comm
 
 # Contribute
 
-In order to update npm package a new release must be created. Any new release must be named @eryxcoop/appyx-comm-{version}
-
+In order to update npm package a new release must be created. Any new release must be named
+@eryxcoop/appyx-comm-{version}
 
 # Work to be done
 
 - [x] Deploy library as a package in npm
 - [ ] Create sample project to show how the package works
 - [x] Add ApiClient
-- [ ] Add tests
+- [x] Add tests
